+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.3.8'
-export const BUILD_DATE = '2026-06-06 16:00'
+export const APP_BUILD = 'v1.4.0'
+export const BUILD_DATE = '2026-06-06 17:30'